Create homepage view and home controller.
Moved the route Application#home to Home#home, where HomeController is new, so that we can bypass authentication for this action. This is necessary so that we can bypass authentication and render the new homepage template for users who are not signed in (previously there was no homepage, and they were just taken to the sign-in page).

The homepage template is in a new view folder 'home', and I've moved the dashboard templates there as well.