Clean up generic controller tests, make more
I've made it so that if you want to use http_request_proc, you are required to pass it in (as opposed to the shared example assuming it exists). I've also tried to move everything away from accessing the lazy/memoized functions via `method(method_name).call` to using alias method. It looks cleaner and I feel like it is more efficient (although I don't know for sure that that is true). I've also created two more generic controller shared examples, and added a the `note` attribute to the `setter_climb_log` model (it made testing easier... and we need it eventually).