Add a little structure to the homepage
Put the sign-in and sign-up forms in a grid; will be side-by-side on all device sizes, except on phone. Also removed the sign-up link from the sign-in form since they are now side by side on the homepage. This also removed it from the plain sign in page, but that is okay since it isn't the landing page anymore, and even on that page, they can still access the sign up page (link is in navbar).