Refine athlete log association with boulder/route
I made a module that you can use to define a type of association where the child is an STI model, and the parent should have only one of the STI types (e.g. either boulder or route). This is used in the parent model, after the `has_one` association declaration, and defines methods for the STI types that you can use to access and build the associated records.