Better comments and error messages for grade enums
closes #42

I decided not to explicitly write out the IDs. If I do it with a loop, it won't be much different than what I have now (in terms of the degree of "future-proofness"). I don't want to list everything out by hand because that is messy and error prone. I decided to just make it more obvious that something went wrong if the IDs are changed.