Add module to generate validations via db cnstrnts
Creates ValidateByReflections module. This module inspects the database constraints and creates corresponding ActiveRecord validations unless one already exists. It is included in ActiveRecord::Base, but you need to call `create_validations` in the model in order to have it run. In the future, we could look at having it happen automatically for each model, or for a set of models based on a configuration file.

A lot of helper modules were created. They are all in concerns, but they don't all extend functionality of ActiveRecord::Base. E.g. one module (ValidationInspector) adds methods to ActiveRecord::ConnectionAdapter::Column and ActiveRecord::Reflection::AssociationReflection so that the column/association can be evaluated to determine if a validation should be created for that column/association.