Just use one seq. in user factory & add init. val.
The initial value helps when we have users already in the database, and we want to open up the console and create more records with FactoryGirl.

The athlete and setter specific factories were removed because it was hard to determine where their sequences should start and we weren't getting significant benefits from them - all they allowed us to do was know that "athlete1@example.com" was defined without having to look it up in the console.